Separate concerns & update version
- Create `Markdown` class to handle all markdown generation
- Move functionality to `Document` class to create instance
