+++ lib/rails-mermaid_erd_markdown/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module MermaidErdMarkdown
-  VERSION = "0.2.0"
+  VERSION = "0.3.0"
 end
